refactor: cleanup tests code
