+++ tests/utils/callCli.js
@@ -2,7 +2,8 @@ const execa = require('execa')
 const cliPath = require('./cliPath')
 
 async function callCli(args, execOptions) {
-  return (await execa(cliPath, args, execOptions)).stdout
+  const { stdout } = await execa(cliPath, args, { windowsHide: true, windowsVerbatimArguments: true, ...execOptions })
+  return stdout
 }
 
 module.exports = callCli
