+++ tests/utils/createLiveTestSite.js
@@ -1,8 +1,27 @@
 const stripAnsi = require('strip-ansi')
 const callCli = require('./callCli')
 
-async function createLiveTestSite(siteName, accountSlug, execOptions) {
-  const cliResponse = await callCli(['sites:create', '--name', siteName, '--account-slug', accountSlug], execOptions)
+function generateSiteName(prefix) {
+  const randomString = Math.random()
+    .toString(36)
+    .replace(/[^a-z]+/g, '')
+    .substr(0, 8)
+  return `${prefix}${randomString}`
+}
+
+async function listAccounts() {
+  return JSON.parse(await callCli(['api', 'listAccountsForUser']))
+}
+
+async function createLiveTestSite(siteName) {
+  console.log(`Creating new site for tests: ${siteName}`)
+  const accounts = await listAccounts()
+  if (!Array.isArray(accounts) || accounts.length <= 0) {
+    throw new Error(`Can't find suitable account to create a site`)
+  }
+  const accountSlug = accounts[0].slug
+  console.log(`Using account ${accountSlug} to create site: ${siteName}`)
+  const cliResponse = await callCli(['sites:create', '--name', siteName, '--account-slug', accountSlug])
 
   const isSiteCreated = /Site Created/.test(cliResponse)
   if (!isSiteCreated) {
@@ -17,4 +36,4 @@ async function createLiveTestSite(siteName, accountSlug, execOptions) {
   return null
 }
 
-module.exports = createLiveTestSite
+module.exports = { generateSiteName, createLiveTestSite }
